requirements.txt
# local package
-e .

# external requirements
awscli
click
coverage
flake8
Sphinx
pywin32==227
naive-bayes==0.1.1
matplotlib-inline==0.1.3
six==1.16.0
tldextract==3.2.0
h5py==3.6.0
decorator==5.1.0
python-dotenv>=0.5.1
pickleshare==0.7.5
seaborn==0.11.2
scipy==1.7.1
numpy==1.20.1
PyYAML==5.4.1
pandas==1.3.3
numexpr==2.7.3
python-math
matplotlib==3.4.3
pywinpty==1.1.4
notebook==6.4.4
pytz==2021.1
sklearn==0.0
pyts==0.12.0
scikit-image==0.19.1
httplib2==0.20.2.
scikit-learn==0.24.2
shap==0.40.0
numba==0.54.1
jupyter==1.0.0
catboost==1.0.4
kafka-python==2.0.2
DateTime==4.4
jupyter-client==7.0.3
requests
logging